requirements.txt
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
torch==2.1.1
matplotlib==3.8.2
torchvision==0.16.1
torchsummary==1.5.1
torchshow==0.5.1
